third_party/blink/web_tests/webmidi/implicit-open-expected.txt

CONSOLE WARNING: line 61: Web MIDI will ask a permission to use even if the sysex is not specified in the MIDIOptions since M77, around September 2019. See https://www.chromestatus.com/feature/5138066234671104 for more details.
Tests MIDIPort implicit open.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS PermissionsHelper.setPermission is defined.
PASS requestMIDIAccess() succeeded with access [object MIDIAccess].
Check state transition for setonmidimessage on closed state.
- check initial state.
PASS port.connection is "closed"
- check port handler.
PASS handler is called with port [object MIDIInput].
PASS eventport.id is "MockInputID-0"
PASS eventport.connection is "open"
- check access handler.
PASS handler is called with port [object MIDIInput].
PASS eventport.id is "MockInputID-0"
PASS eventport.connection is "open"
- reset input device
Check state transition for addeventlistener on closed state.
- check initial state.
PASS port.connection is "closed"
- check port handler.
PASS handler is called with port [object MIDIInput].
PASS eventport.id is "MockInputID-0"
PASS eventport.connection is "open"
- check access handler.
PASS handler is called with port [object MIDIInput].
PASS eventport.id is "MockInputID-0"
PASS eventport.connection is "open"
Check state transition for send on closed state.
- check initial state.
PASS port.connection is "closed"
- check port handler.
PASS handler is called with port [object MIDIOutput].
PASS eventport.id is "MockOutputID-0"
PASS eventport.connection is "open"
- check access handler.
PASS handler is called with port [object MIDIOutput].
PASS eventport.id is "MockOutputID-0"
PASS eventport.connection is "open"
PASS successfullyParsed is true

TEST COMPLETE
